+++ Lab2/bootloader/include/reboot.h
@@ -1,0 +1,8 @@
+#ifndef REBOOT_H
+#define REBOOT_H
+
+extern int EXIT;
+void reset(unsigned int);
+void cancel_reset(void);
+
+#endif /* REBOOT_H */
